refactor(renderer): extract the bloom chain into postfx_chain.rs
Render-graph migration cluster 5 prep: the Karis-thresholded
downsample + additive-upsample chain becomes record_bloom_chain; the
SSR/SSGI composite-input view bindings move below the last &mut-self
pass recording (compose reads them later either way). Goldens
pixel-identical. mod.rs ratchets 12913 -> 12787.
